fix: add hasExportsSourceField util
