+++ packages/monorepo-utils/src/project-utils/index.ts
@@ -1,3 +1,2 @@
 export * from './getDependentProjects';
-export * from './getExportsSourceDirs';
 export * from './filter';
